Домашнее задание     Обработчик json-ов
