+++ hw08-gradle/src/main/java/ru/otus/dataprocessor/Parser.java
@@ -1,0 +1,8 @@
+package ru.otus.dataprocessor;
+
+import java.io.IOException;
+import java.io.InputStream;
+
+public interface Parser<M>{
+    M parse(InputStream inputStream) throws IOException;
+}
