+++ hw08-gradle/src/main/java/ru/otus/dataprocessor/ParserFactory.java
@@ -1,0 +1,11 @@
+package ru.otus.dataprocessor;
+
+import ru.otus.model.Measurement;
+
+import java.util.List;
+
+public class ParserFactory {
+    public static Parser<List<Measurement>> getMeasurementParser() {
+        return new MeasurementParser();
+    }
+}
