+++ hw08-gradle/src/test/java/ru/otus/dataprocessor/ProcessorTest.java
@@ -1,7 +1,5 @@
 package ru.otus.dataprocessor;
 
-import org.assertj.core.api.AssertionsForClassTypes;
-import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.io.TempDir;
